added update customer profile and block feature implemented

+++ database/migrations/2016_06_20_085355_create_user_details_table.php
@@ -17,9 +17,9 @@ class CreateUserDetailsTable extends Migration
             $table->integer('user_id');
             $table->string('name');
             $table->longText('address');
-            $table->integer('personal_ph');
-            $table->integer('cell_phone')->nullable();
-            $table->integer('off_phone')->nullable();
+            $table->bigInteger('personal_ph');
+            $table->bigInteger('cell_phone')->nullable();
+            $table->bigInteger('off_phone')->nullable();
             $table->string('spcl_instructions')->nullable();
             $table->string('driving_instructions')->nullable();
             $table->tinyInteger('payment_status')->comment = "1->paid , 0->pending";
